feat(cloud-provisioning): Refactored code:
- Moved webclient package from internal to pkg
- Moved CloudProvider struct from cloudproviders package to domain package
- Updated imports

Closes VC-33251
