+++ pkg/webclient/cloudproviders/cloudproviders.gen.go
